backend/requirements.txt
# ─── Web framework ───
fastapi==0.115.0
uvicorn[standard]==0.32.0

# ─── Data validation & config ───
pydantic==2.9.2
pydantic-settings==2.6.0
PyYAML==6.0.2

# ─── Storage ───
aiosqlite==0.20.0

# ─── HTTP client (Step 2) ───
httpx==0.27.2

# ─── Scheduler (Step 2) ───
APScheduler==3.10.4

# ─── Resilience (Step 2) ───
tenacity==9.0.0

# ─── Misc ───
python-dotenv==1.0.1
orjson==3.10.7
